+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/KeyboardShortcutsPanel.java
@@ -149,6 +149,7 @@ final class KeyboardShortcutsPanel extends JPanel {
 		return gridLayoutPanel(0, 2)
 						.addAll(new JLabel(MESSAGES.getString("previous_next_operator")), new JLabel(CTRL + UP_DOWN))
 						.addAll(new JLabel(MESSAGES.getString("enable_disable_condition")), new JLabel(CTRL + ENTER))
+						.addAll(new JLabel(MESSAGES.getString("clear_condition")), new JLabel(CTRL + SHIFT + ENTER))
 						.addAll(new JLabel(MESSAGES.getString("refresh_table_data")), new JLabel(ENTER))
 						.border(createTitledBorder(MESSAGES.getString("condition_panel")))
 						.build();